Création du iitinaryController
